// src/ecom-sadec-api-client/types.gen.ts
export type HandOverStatusEnum = 0 | 1;
// ...
export type LegalStatusEnum = 1 | 2 | 3 | 4;

export type ListingFavoriteInputDto = {
  listingId?: string | null;
  isFavorite?: boolean;
};

export type ListingFavoriteSearchQueryInputDto = {
  from?: number;
  size?: number;
  keyword?: string | null;
  type?: ListingTypeEnum;
};

export type ListingLeaseTermEnum = 1 | 2 | 3 | 4;

export type ListingMarkAsVipInputDto = {
  id?: string | null;
  priority?: ListingPackageTypeEnum;
};

export type ListingPackageEnum = 1 | 2 | 3 | 4;

export type ListingPackageStatusEnum = 0 | 1;

export type ListingPackageTypeEnum = 1 | 2 | 3;

export type ListingPushStatusEnum = 0 | 1;

export type ListingSearchQueryDto = {
  type?: ListingTypeEnum;
  projectIds?: Array<string> | null;
  categoryIds?: Array<string> | null;
  viewsIds?: Array<string> | null;
  bedrooms?: Array<number> | null;
  bathrooms?: Array<number> | null;
  fromPriceUsd?: number | null;
  toPriceUsd?: number | null;
  fromPriceVnd?: number | null;
  toPriceVnd?: number | null;
  priorityStatus?: ListPriorityStatusEnum;
  fromSize?: number | null;
  toSize?: number | null;
  province?: string | null;
  district?: string | null;
  ward?: string | null;
  keyword?: string | null;
  currentLocation?: SearchLocationInputDto;
  agencyId?: string | null;
  places?: Array<SearchPlaceLocationInputDto> | null;
  inDoorAmenities?: Array<string> | null;
  outDoorAmenities?: Array<string> | null;
  handOverStatus?: HandOverStatusEnum;
  legalStatus?: Array<LegalStatusEnum> | null;
  leaseTerm?: ListingLeaseTermEnum;
  isPetAllowance?: boolean | null;
  interiorIds?: Array<string> | null;
  orderByCreatedTime?: boolean | null;
  contentLanguage?: Array<string> | null;
  fromMoveDate?: string | null;
  toMoveDate?: string | null;
  posterBrokerId?: string | null;
  from?: number;
  size?: number;
};

export type ListingStatusEnum = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 99 | 100;

export type ListingTypeEnum = 0 | 1;

export type ListingUpdateActualPrice = {
  listingId?: string | null;
  actualPrice?: number | null;
};

export type ListingUpdateStatusDto = {
  id: string;
  status: ListingStatusEnum;
  message?: string | null;
};

export type ListPriorityStatusEnum = 1 | 2 | 3;
// ...
export type SearchLocationInputDto = {
  lat?: number;
  lng?: number;
};

export type SearchNotificationRequest = {
  from?: number;
  size?: number;
  method?: NotificationMethod;
  notiType?: NotificationTypeEnum;
};

export type SearchPlaceLocationInputDto = {
  distance?: number;
  id?: string | null;
};